# backend/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  app.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  models.Invoice:
    properties:
      created_at:
        type: string
      id:
        type: integer
      nodes:
        items:
          $ref: '#/definitions/models.NodeItem'
        type: array
      tax:
        type: number
      total:
        type: number
      user_id:
        type: integer
    required:
    - user_id
    type: object
  models.NodeItem:
    properties:
      contract_id:
        type: integer
      cost:
        type: number
      id:
        type: integer
      invoice_id:
        type: integer
      node_id:
        type: integer
      period:
        type: number
      rent_created_at:
        type: string
    type: object
...
info:
  contact: {}
  description: API documentation for Mycelium Cloud.
  title: Mycelium Cloud API
  version: "1.0"
paths:
...
  /invoices:
    get:
      consumes:
      - application/json
      description: Returns a list of all invoices
      operationId: get-all-invoices
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Invoice'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - AdminMiddleware: []
      summary: Get all invoices
      tags:
      - admin
...
  /user/invoice:
    get:
      consumes:
      - application/json
      description: Returns a list of invoices for a user
      operationId: get-invoices
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Invoice'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - UserMiddleware: []
      summary: Get invoices
      tags:
      - invoices
  /user/invoice/{invoice_id}:
    get:
      consumes:
      - application/json
      description: Downloads an invoice by ID
      operationId: download-invoice
      parameters:
      - description: Invoice ID
        in: path
        name: invoice_id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Invoice is not found
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - UserMiddleware: []
      summary: Download invoice
      tags:
      - invoices
...
swagger: "2.0"